Started on the layout and added boxes for Current and next job
Started on the layout for the driver view and have added a table with 3 boxes one for current job one for next job and one for additional info, like arrival time etc. Also swapped around the positions for the buttons so that the current job only have a finish button and the next job have an accept and decline button for job requests. 

Idea that the "accept" button changes to "next job" after one has pressed it once so that you get an easy method to transition to your next job.