Added button reciept and box for customer number
Branch from Dimitris driver code, added the box where customer number should be added. The box doesn't have any functionality to it since it should only be used in the simulation. If the box should only be visible when a "färdtjänst" job is ongoing then a if statement needs to be added. 
The button for receipt is also purely for the simulation and doesn't have any other functionality than being pressed.