Added box for customer number and print receipt button
Added the html and the styling for the input box for customer number aswell as the button print receipt. These should not have any back end and are simply there for us to show the goals in our simulation. As of now the print receipt shows on all trips and not only on the "färdtjänst".